Migrate sql/peers to db/peers as pg-promise repo
