+++ test/unit/modules/peers.js
@@ -33,7 +33,10 @@ describe('peers', function () {
 
 	before(function () {
 		dbMock = {
-			any: sinon.stub().resolves()
+			any: sinon.stub().resolves(),
+			peers: {
+				list: sinon.stub().resolves()
+			}
 		};
 		PeersRewired = rewire('../../../modules/peers');
 		peersLogicMock = {
